// src/types.ts
export interface PortfolioItem {
  id: string;
  title: string;
  titleMr?: string;
  category: 'logo' | 'social' | 'packaging' | 'brochure' | 'outdoor' | 'stationery' | 'branding' | string;
  categoryLabel: string;
  categoryLabelMr?: string;
  client?: string;
  city?: string;
  cityMr?: string;
  image: string;
  aspectRatio?: 'square' | 'portrait' | 'landscape';
  description: string;
  descriptionMr?: string;
  tags?: string[];
  deliverables?: string[];
  deliverablesMr?: string[];
  badge?: string;
  badgeMr?: string;
}